[Feat]: UModel: add security when remove model

+++ Umodel.pas
@@ -106,12 +106,17 @@ uses
 procedure TFModel.AModelDeleteExecute(Sender: TObject);
 var
   id : Integer;
-  query : string;
+  query, name : string;
 begin
   id:=DSModel.DataSet.FieldByName('id').AsInteger;
-  query := 'DELETE FROM model WHERE id=:id';
-  DMDatabase.dbConnection.ExecSQL(query,[id]);
-  reloadValues;
+  name:=DSModel.DataSet.FieldByName('name').AsString;
+  if MessageDlg('Voulez vous supprimer ' + name + ' ?', TMsgDlgType.mtWarning,
+    [TMsgDlgBtn.mbYes, TMsgDlgBtn.mbNo], 0, TMsgDlgBtn.mbNo) = mrYes then
+  begin
+    query := 'DELETE FROM model WHERE id=:id';
+    DMDatabase.dbConnection.ExecSQL(query,[id]);
+    reloadValues;
+  end;
 end;
 
 procedure TFModel.AModelShowDetailsExecute(Sender: TObject);
